+++ aikido/include/aikido/planner/ompl/AIKIDOStateSampler.hpp
@@ -2,7 +2,7 @@
 #define AIKIDO_OMPL_AIKIDOSTATESAMPLER_HPP_
 
 #include <ompl/base/StateSampler.h>
-#include "../constraint/Sampleable.hpp"
+#include "../../constraint/Sampleable.hpp"
 
 namespace aikido {
 namespace ompl {
