+++ aikido/src/CMakeLists.txt
@@ -6,4 +6,3 @@ add_subdirectory("distance")   # Dependencies: statespace
 add_subdirectory("trajectory") # Dependencies: statespace
 add_subdirectory("constraint") # Dependencies: statespace, util
 add_subdirectory("planner")    # Dependencies: external, statespace, trajectory
-add_subdirectory("ompl")       # Dependencies: constraint, statespace, trajectory
