+++ aikido/src/planner/CMakeLists.txt
@@ -1,5 +1,9 @@
 add_subdirectory("parabolic")
 
+if(${OMPL_FOUND})
+  add_subdirectory("ompl")
+endif()
+
 aikido_add_library("${PROJECT_NAME}_planner" SnapPlanner.cpp)
 target_link_libraries("${PROJECT_NAME}_planner"
   "${PROJECT_NAME}_constraint"
