+++ aikido/src/planner/ompl/AIKIDOGeometricStateSpace.cpp
@@ -1,5 +1,5 @@
-#include <aikido/ompl/AIKIDOGeometricStateSpace.hpp>
-#include <aikido/ompl/AIKIDOStateSampler.hpp>
+#include <aikido/planner/ompl/AIKIDOGeometricStateSpace.hpp>
+#include <aikido/planner/ompl/AIKIDOStateSampler.hpp>
 #include <aikido/constraint/Sampleable.hpp>
 #include <dart/common/StlHelpers.h>
 #include <boost/make_shared.hpp>
