+++ aikido/src/planner/ompl/CMakeLists.txt
@@ -1,11 +1,13 @@
-aikido_add_library("${PROJECT_NAME}_ompl"
+include_directories(SYSTEM ${OMPL_INCLUDE_DIRS})
+
+aikido_add_library("${PROJECT_NAME}_planner_ompl"
   AIKIDOGeometricStateSpace.cpp
   AIKIDOStateSampler.cpp
   AIKIDOStateValidityChecker.cpp
   OMPLPlanner.cpp
   GoalRegion.cpp
 )
-target_link_libraries("${PROJECT_NAME}_ompl"
+target_link_libraries("${PROJECT_NAME}_planner_ompl"
   "${PROJECT_NAME}_constraint"
   "${PROJECT_NAME}_distance"
   "${PROJECT_NAME}_statespace"
