+++ aikido/tests/CMakeLists.txt
@@ -7,7 +7,6 @@ include_directories("${CMAKE_CURRENT_SOURCE_DIR}")
 
 add_subdirectory("constraint")
 add_subdirectory("distance")
-add_subdirectory("ompl")
 add_subdirectory("planner")
 add_subdirectory("statespace")
 add_subdirectory("trajectory")
