+++ aikido/tests/planner/CMakeLists.txt
@@ -1,8 +1,13 @@
 add_subdirectory("parabolic")
 
+if(${OMPL_FOUND})
+  add_subdirectory("ompl")
+endif()
+
 aikido_add_test(test_SnapPlanner test_SnapPlanner.cpp)
 target_link_libraries(test_SnapPlanner
   "${PROJECT_NAME}_constraint"
   "${PROJECT_NAME}_distance"
   "${PROJECT_NAME}_trajectory"
   "${PROJECT_NAME}_planner")
+
